Some improvments in perlin alghoritm
