+++ src/PerlinsMap.cpp
@@ -1,10 +1,9 @@
 #include "../include/PerlinsMap.hpp"
 
-#include <glm/glm.hpp>
 #include <glm/gtc/noise.hpp>
 #include <glm/gtx/euler_angles.hpp>
 #include <cmath>
-#include <fstream>
+#include "../include/header_only/BmpGenerator.hpp"
 
 rcn3d::PerlinsMap::PerlinsMap() {
 
@@ -14,30 +13,43 @@ rcn3d::PerlinsMap::~PerlinsMap() {
 
 }
 
-void rcn3d::PerlinsMap::generateMap(int width, int height, float scale) {
+void rcn3d::PerlinsMap::generateMap(unsigned int w, unsigned int h, float s, int o, float p, float l) {
 
-    this->width = width;
-    this->height = height;
+    width = w;
+    height = h;
+	scale = s;
 
-//    std::ofstream outf("ddd.ppm", std::ios::out | std::ios::binary);
-//    outf << "P6\n" << width << " " << height << "\n255\n";
+	if(scale <= 0) {
+		scale = 0.0001f;
+	}
 
-    Canvas can;
+	rcn3d::Canvas can;
     can.make(width, height);
 
-    for(int i = 0; i < height; i++) {
-        for (int j = 0; j < width; j++) {
-            float sampleX = j / scale;
-            float sampleY = i / scale;
+    for(int y = 0; y < height; y++) {
+        for (int x = 0; x < width; x++) {
 
-            float perlinValue = glm::perlin(glm::vec2(sampleX, sampleY)) + 1.0f;
-            perlinValue = fmod(perlinValue, 1.0f);
-            float lerpValue = glm::lerp(0.0f, 255.0f, perlinValue);
-            can.setColor(j, i, lerpValue, lerpValue, lerpValue);
+			float amplitude = 1.0f;
+			float frequency = 1.0f;
+			float noiseHeight = 0.0f;
 
-            mapData.push_back(perlinValue);
+			for(int i = 0; i < o; i++) {
+				float sampleX = x / scale * frequency;
+				float sampleY = y / scale * frequency;
+
+				// Little hack since glm::perlin return number between -1.0f and 1.0f
+				float perlinValue = glm::perlin(glm::vec2(sampleX, sampleY)) + 1.0f;
+				perlinValue = static_cast<float>(fmod(perlinValue, 1.0f));
+				noiseHeight += perlinValue * amplitude;
+
+				amplitude *= p;
+				frequency *= l;
+			}
+
+			int lerpValue = static_cast<int>(glm::lerp(0.0f, 255.0f, noiseHeight));
+			can.setColor(x, y, lerpValue, lerpValue, lerpValue);
+			mapData.push_back(lerpValue);
         }
     }
-    can.save("perlin100x100.bmp");
-
+    can.save("perlin_noise_new.bmp");
 }
